[State Reconciliation] Split into several files.
