+++ Artsy/View_Controllers/Live_Auctions/LiveAuctionSalesPerson.swift
@@ -4,15 +4,19 @@ import Interstellar
 /// Something to pretend to either be a network model or whatever
 /// for now it can just parse the embedded json, and move it to obj-c when we're doing real networking
 
+// TODO: This is getting pretty heavy, split up or scale down.
 protocol LiveAuctionsSalesPersonType {
     var currentIndexSignal: Signal<Int> { get }
+    var newLotsSignal: Signal<[LiveAuctionLotViewModel]> { get }
+    var saleSignal: Signal<LiveAuctionViewModel> { get }
+    var currentLotSignal: Signal<LiveAuctionLotViewModel> { get }
     var auctionViewModel: LiveAuctionViewModel? { get }
     var pageControllerDelegate: LiveAuctionPageControllerDelegate! { get }
 
     func lotViewModelForIndex(index: Int) -> LiveAuctionLotViewModel?
     func lotViewModelRelativeToShowingIndex(offset: Int) -> LiveAuctionLotViewModel?
+
     var lotCount: Int { get }
-    var updatedState: Signal<LiveAuctionsSalesPersonType> { get }
 }
 
 class LiveAuctionsSalesPerson:  NSObject, LiveAuctionsSalesPersonType {
@@ -26,6 +30,7 @@ class LiveAuctionsSalesPerson:  NSObject, LiveAuctionsSalesPersonType {
     private var sale: LiveAuctionViewModel?
     private let stateManager: LiveAuctionStateManager
 
+    // TODO: What does "current index" mean? It should be what the sale is on, not what the user is looking at.
     var currentIndexSignal = Signal<Int>()
 
     func lotViewModelRelativeToShowingIndex(offset: Int) -> LiveAuctionLotViewModel? {
@@ -68,6 +73,23 @@ class LiveAuctionsSalesPerson:  NSObject, LiveAuctionsSalesPersonType {
 
     var lotCount: Int {
         return auctionViewModel?.lotCount ?? 0
+    }
+}
+
+private typealias ComputedProperties = LiveAuctionsSalesPerson
+extension ComputedProperties {
+    var currentLotSignal: Signal<LiveAuctionLotViewModel> {
+        return stateManager.currentLotSignal
+    }
+
+    // TODO: React to this somehow
+    var newLotsSignal: Signal<[LiveAuctionLotViewModel]> {
+        return stateManager.newLotsSignal
+    }
+
+    // TODO: React to this somehow in the VC
+    var saleSignal: Signal<LiveAuctionViewModel> {
+        return stateManager.saleSignal
     }
 }
